#5 get_data_from_hdf5 file created
